Implement BFS without paths or targets.
This implements the first, simplest version of bread-first-search based on the Python code in [Introduction to A*](http://www.redblobgames.com/pathfinding/a-star/introduction.html). This doesn't actually check to see if you've found the target state, and it doesn't keep track of how you get from one state to another, but it does manage to search the _entire_ 8-puzzle space in just over 5 seconds on my iMac. Yowza!