Increase heap size for leiningens
I occasionally ran into problems here the search process would run out of heap, so I bumped that up to a GB. This not only fixed the problem, but appears to have substantially sped up the program.

While I was there I also explicitly added midje as a plugin in the project file.