Implement variant of `num-wrong` that doesn't count the blank
It seems that the simple version of `num-wrong` keeps trying to bring the blank back to the top-left corner, and that probably isn't good for trying to solve the problem. This adds `num-non-blank-wrong` that just counts the number of non-blank tiles that aren't in the right place. This does in fact lead to searching fewer states and (slightly) shorter solutions.