Explore lazy search paper ideas
This implements a number of different ideas,
including some from my (very old) tech report
on lazy evaluation and search. It turns out
that you can't really push things quite that
hard because `map`s in Clojure aren't lazy
and you get a stack overflow just building
the tree.

I think there are ways to deal with this,
but I really don't have time to deal with it
now.